refactor: period-rules を (features)/_shared へ昇格し中立型化
共通売単価 _data/period-rules を (features)/_shared へ移動し、入力を集約 DTO
依存の status から中立型 PeriodStatus（future|active|expired）へ切り替えた。
原価(#503)・得意先別販売単価(#513) が同じ status を持ち rule of three が見えて
いるため、複製ではなく昇格を選択（#501 の formatYen 昇格と同じ判断枠）。

authorityFor は集約非依存の純粋述語で、各編集読みモデルの status を structural
typing でそのまま渡せる。集約 DTO 同士の結合は生まない。将来 権限ルールが集約間で
分岐したらフォークすれば十分（可逆）。
